+++ lib/src/models/audio_filter.dart
@@ -1,3 +1,7 @@
+// Copyright © 2026 & onwards, Alessandro Di Ronza <ales.drnz@gmail.com>.
+// All rights reserved.
+// Use of this source code is governed by MIT license that can be found in the LICENSE file.
+
 /// An audio filter applied to the mpv filter graph (`--af`).
 ///
 /// Filters use mpv's libavfilter notation. Named constructors cover the most