+++ lib/src/models/audio_params.dart
@@ -1,3 +1,7 @@
+// Copyright © 2026 & onwards, Alessandro Di Ronza <ales.drnz@gmail.com>.
+// All rights reserved.
+// Use of this source code is governed by MIT license that can be found in the LICENSE file.
+
 /// Audio format parameters as reported by the mpv audio output pipeline.
 class AudioParams {
   /// Sample format string (e.g. `"floatp"`, `"s16"`).