+++ lib/src/models/gapless_mode.dart
@@ -1,3 +1,7 @@
+// Copyright © 2026 & onwards, Alessandro Di Ronza <ales.drnz@gmail.com>.
+// All rights reserved.
+// Use of this source code is governed by MIT license that can be found in the LICENSE file.
+
 /// Modes for gapless audio transitions.
 enum GaplessMode {
   /// Disable gapless playback.