+++ lib/src/models/playlist.dart
@@ -1,3 +1,7 @@
+// Copyright © 2026 & onwards, Alessandro Di Ronza <ales.drnz@gmail.com>.
+// All rights reserved.
+// Use of this source code is governed by MIT license that can be found in the LICENSE file.
+
 import 'package:mpv_audio_kit/src/models/media.dart';
 
 /// The repeat / loop mode for the player's playlist.