+++ lib/src/utils/android_helper.dart
@@ -2,6 +2,7 @@
 // Copyright © 2021 & onwards, Hitesh Kumar Saini <saini123hitesh@gmail.com>.
 // All rights reserved.
 // Use of this source code is governed by MIT license that can be found in the LICENSE file.
+
 import 'dart:io';
 import 'package:flutter/services.dart';
 import 'package:flutter/foundation.dart';
@@ -19,34 +20,38 @@ abstract class AndroidHelper {
         return await _copyAssetToCache(uri);
       }
       if (Platform.isAndroid && uri.startsWith('content://')) {
-        final fd = await _channel.invokeMethod<int>('openFileDescriptor', {'uri': uri});
+        final fd = await _channel
+            .invokeMethod<int>('openFileDescriptor', {'uri': uri});
         if (fd != null && fd > 0) {
           return 'fd://$fd';
         }
       }
     } catch (e) {
-      debugPrint('mpv_audio_kit: AndroidHelper.normalizeUri error for $uri: $e');
+      debugPrint(
+          'mpv_audio_kit: AndroidHelper.normalizeUri error for $uri: $e');
     }
     return uri;
   }
 
   static Future<String> _copyAssetToCache(String uri) async {
     if (_assetCache.containsKey(uri)) return _assetCache[uri]!;
-    
+
     // Extract the raw path inside the asset bundle
     String assetPath = uri.substring('asset://'.length);
     if (assetPath.startsWith('/')) {
       assetPath = assetPath.substring(1);
     }
-    
+
     final data = await rootBundle.load(assetPath);
-    
+
     // Sanitize path for use as a filesystem path
-    final safeName = assetPath.replaceAll(Platform.pathSeparator, '_').replaceAll('/', '_');
-    final file = File('${Directory.systemTemp.path}${Platform.pathSeparator}mpv_asset_$safeName');
-    
+    final safeName =
+        assetPath.replaceAll(Platform.pathSeparator, '_').replaceAll('/', '_');
+    final file = File(
+        '${Directory.systemTemp.path}${Platform.pathSeparator}mpv_asset_$safeName');
+
     await file.writeAsBytes(data.buffer.asUint8List(), flush: true);
-    
+
     _assetCache[uri] = file.path;
     return file.path;
   }